chat-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
pgvector>=0.2.0
sqlalchemy>=2.0.35
alembic>=1.14.0
psycopg2-binary>=2.9.5
redis>=5.2.0
pydantic>=2.7.4
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.28.0
python-dotenv>=1.1.1
openai>=1.100.1
jinja2>=3.1.0
loguru>=0.7.0
structlog>=24.0.0